fix The argument type 'num' can't be assigned to the parameter type 'double'.

+++ speech_to_text/lib/speech_to_text_web.dart
@@ -238,7 +238,7 @@ class SpeechToTextPlugin extends SpeechToTextPlatform {
         final transcript = alt.transcript;
         final confidence = alt.confidence;
 
-        balanced.add(resultIndex, transcript, confidence);
+        balanced.add(resultIndex, transcript, confidence.toDouble());
       }
       ++resultIndex;
     }
